+++ crates/bevy_core_pipeline/src/skybox/mod.rs
@@ -33,7 +33,7 @@ use bevy_utils::default;
 use prepass::SkyboxPrepassPipeline;
 
 use crate::{
-    core_3d::CORE_3D_DEPTH_FORMAT, prepass::PreviousViewUniforms,
+    prepass::PreviousViewUniforms,
     skybox::prepass::init_skybox_prepass_pipeline,
 };
