+++ crates/bevy_pbr/src/lightmap/mod.rs
@@ -50,12 +50,7 @@ use bevy_math::{uvec2, vec4, Rect, UVec2};
 use bevy_platform::collections::HashSet;
 use bevy_reflect::{std_traits::ReflectDefault, Reflect};
 use bevy_render::{
-    render_asset::RenderAssets,
-    render_resource::{Sampler, TextureView, WgpuSampler, WgpuTextureView},
-    renderer::RenderAdapter,
-    sync_world::MainEntity,
-    texture::{FallbackImage, GpuImage},
-    Extract, ExtractSchedule, RenderApp, RenderStartup,
+    mesh::lightmap::{LightmapSlabIndex, LightmapSlotIndex, RenderLightmap, RenderLightmapsU}, render_asset::RenderAssets, render_resource::{Sampler, TextureView, WgpuSampler, WgpuTextureView}, renderer::RenderAdapter, sync_world::MainEntity, texture::{FallbackImage, GpuImage}, Extract, ExtractSchedule, RenderApp, RenderStartup
 };
 use bevy_render::{renderer::RenderDevice, sync_world::MainEntityHashMap};
 use bevy_shader::load_shader_library;
@@ -107,53 +102,11 @@ pub struct Lightmap {
     pub bicubic_sampling: bool,
 }
 
-/// Lightmap data stored in the render world.
-///
-/// There is one of these per visible lightmapped mesh instance.
-#[derive(Debug)]
-pub(crate) struct RenderLightmap {
-    /// The rectangle within the lightmap texture that the UVs are relative to.
-    ///
-    /// The top left coordinate is the `min` part of the rect, and the bottom
-    /// right coordinate is the `max` part of the rect. The rect ranges from (0,
-    /// 0) to (1, 1).
-    pub(crate) uv_rect: Rect,
-
-    /// The index of the slab (i.e. binding array) in which the lightmap is
-    /// located.
-    pub(crate) slab_index: LightmapSlabIndex,
-
-    /// The index of the slot (i.e. element within the binding array) in which
-    /// the lightmap is located.
-    ///
-    /// If bindless lightmaps aren't in use, this will be 0.
-    pub(crate) slot_index: LightmapSlotIndex,
-
-    // Whether or not bicubic sampling should be used for this lightmap.
-    pub(crate) bicubic_sampling: bool,
-}
-
-/// Stores data for all lightmaps in the render world.
-///
-/// This is cleared and repopulated each frame during the `extract_lightmaps`
-/// system.
+/// Corresponds to `RenderLightmapsU`
 #[derive(Resource)]
-pub struct RenderLightmaps {
-    /// The mapping from every lightmapped entity to its lightmap info.
-    ///
-    /// Entities without lightmaps, or for which the mesh or lightmap isn't
-    /// loaded, won't have entries in this table.
-    pub(crate) render_lightmaps: MainEntityHashMap<RenderLightmap>,
-
+pub struct RenderLightmapsL {
     /// The slabs (binding arrays) containing the lightmaps.
-    pub(crate) slabs: Vec<LightmapSlab>,
-
-    free_slabs: FixedBitSet,
-
-    pending_lightmaps: HashSet<(LightmapSlabIndex, LightmapSlotIndex)>,
-
-    /// Whether bindless textures are supported on this platform.
-    pub(crate) bindless_supported: bool,
+    pub slabs: Vec<LightmapSlab>,
 }
 
 /// A binding array that contains lightmaps.
@@ -167,20 +120,7 @@ pub struct LightmapSlab {
 
 struct AllocatedLightmap {
     gpu_image: GpuImage,
-    // This will only be present if the lightmap is allocated but not loaded.
-    asset_id: Option<AssetId<Image>>,
 }
-
-/// The index of the slab (binding array) in which a lightmap is located.
-#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug, Deref, DerefMut)]
-#[repr(transparent)]
-pub struct LightmapSlabIndex(pub(crate) NonMaxU32);
-
-/// The index of the slot (element within the binding array) in the slab in
-/// which a lightmap is located.
-#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug, Deref, DerefMut)]
-#[repr(transparent)]
-pub struct LightmapSlotIndex(pub(crate) NonMaxU16);
 
 impl Plugin for LightmapPlugin {
     fn build(&self, app: &mut App) {
@@ -201,7 +141,8 @@ impl Plugin for LightmapPlugin {
 /// Extracts all lightmaps from the scene and populates the [`RenderLightmaps`]
 /// resource.
 fn extract_lightmaps(
-    render_lightmaps: ResMut<RenderLightmaps>,
+    render_lightmaps_u: ResMut<RenderLightmapsU>,
+    mut render_lightmaps_l: ResMut<RenderLightmapsL>,
     changed_lightmaps_query: Extract<
         Query<
             (Entity, &ViewVisibility, &Lightmap),
@@ -212,11 +153,12 @@ fn extract_lightmaps(
     images: Res<RenderAssets<GpuImage>>,
     fallback_images: Res<FallbackImage>,
 ) {
-    let render_lightmaps = render_lightmaps.into_inner();
+    let render_lightmaps_u2 = render_lightmaps_u.into_inner();
+    let render_lightmaps_l2 = render_lightmaps_l.into_inner();
 
     // Loop over each entity.
     for (entity, view_visibility, lightmap) in changed_lightmaps_query.iter() {
-        if render_lightmaps
+        if render_lightmaps_u2
             .render_lightmaps
             .contains_key(&MainEntity::from(entity))
         {
@@ -229,8 +171,8 @@ fn extract_lightmaps(
         }
 
         let (slab_index, slot_index) =
-            render_lightmaps.allocate(&fallback_images, lightmap.image.id());
-        render_lightmaps.render_lightmaps.insert(
+            render_lightmaps_l2.allocate(render_lightmaps_u2, &fallback_images, lightmap.image.id());
+        render_lightmaps_u2.render_lightmaps.insert(
             entity.into(),
             RenderLightmap::new(
                 lightmap.uv_rect,
@@ -240,7 +182,7 @@ fn extract_lightmaps(
             ),
         );
 
-        render_lightmaps
+        render_lightmaps_u2
             .pending_lightmaps
             .insert((slab_index, slot_index));
     }
@@ -254,72 +196,39 @@ fn extract_lightmaps(
             slab_index,
             slot_index,
             ..
-        }) = render_lightmaps
+        }) = render_lightmaps_u2
             .render_lightmaps
             .remove(&MainEntity::from(entity))
         else {
             continue;
         };
 
-        render_lightmaps.remove(&fallback_images, slab_index, slot_index);
-        render_lightmaps
+        // render_lightmaps_l.remove(render_lightmaps_u2, &fallback_images, slab_index, slot_index);
+        render_lightmaps_u2
             .pending_lightmaps
             .remove(&(slab_index, slot_index));
     }
 
-    render_lightmaps
+    render_lightmaps_u2
         .pending_lightmaps
         .retain(|&(slab_index, slot_index)| {
-            let Some(asset_id) = render_lightmaps.slabs[usize::from(slab_index)].lightmaps
-                [usize::from(slot_index)]
-            .asset_id
-            else {
-                error!(
-                    "Allocated lightmap should have been removed from `pending_lightmaps` by now"
-                );
-                return false;
-            };
+            return false;
+            // let asset_id = render_lightmaps_u2.slabs[usize::from(slab_index)].lightmaps
+            //     [usize::from(slot_index)]
+            // .asset_id
+            // else {
+            //     error!(
+            //         "Allocated lightmap should have been removed from `pending_lightmaps` by now"
+            //     );
+            //     return false;
+            // };
 
-            let Some(gpu_image) = images.get(asset_id) else {
-                return true;
-            };
-            render_lightmaps.slabs[usize::from(slab_index)].insert(slot_index, gpu_image.clone());
-            false
+            // let Some(gpu_image) = images.get(asset_id) else {
+            //     return true;
+            // };
+            // render_lightmaps_l.slabs[usize::from(slab_index)].insert(slot_index, gpu_image.clone());
+            // false
         });
-}
-
-impl RenderLightmap {
-    /// Creates a new lightmap from a texture, a UV rect, and a slab and slot
-    /// index pair.
-    fn new(
-        uv_rect: Rect,
-        slab_index: LightmapSlabIndex,
-        slot_index: LightmapSlotIndex,
-        bicubic_sampling: bool,
-    ) -> Self {
-        Self {
-            uv_rect,
-            slab_index,
-            slot_index,
-            bicubic_sampling,
-        }
-    }
-}
-
-/// Packs the lightmap UV rect into 64 bits (4 16-bit unsigned integers).
-pub(crate) fn pack_lightmap_uv_rect(maybe_rect: Option<Rect>) -> UVec2 {
-    match maybe_rect {
-        Some(rect) => {
-            let rect_uvec4 = (vec4(rect.min.x, rect.min.y, rect.max.x, rect.max.y) * 65535.0)
-                .round()
-                .as_uvec4();
-            uvec2(
-                rect_uvec4.x | (rect_uvec4.y << 16),
-                rect_uvec4.z | (rect_uvec4.w << 16),
-            )
-        }
-        None => UVec2::ZERO,
-    }
 }
 
 impl Default for Lightmap {
@@ -339,58 +248,13 @@ pub fn init_render_lightmaps(
 ) {
     let bindless_supported = binding_arrays_are_usable(&render_device, &render_adapter);
 
-    commands.insert_resource(RenderLightmaps {
+    commands.insert_resource(RenderLightmapsU {
         render_lightmaps: default(),
         slabs: vec![],
         free_slabs: FixedBitSet::new(),
         pending_lightmaps: default(),
         bindless_supported,
     });
-}
-
-impl RenderLightmaps {
-    /// Creates a new slab, appends it to the end of the list, and returns its
-    /// slab index.
-    fn create_slab(&mut self, fallback_images: &FallbackImage) -> LightmapSlabIndex {
-        let slab_index = LightmapSlabIndex::from(self.slabs.len());
-        self.free_slabs.grow_and_insert(slab_index.into());
-        self.slabs
-            .push(LightmapSlab::new(fallback_images, self.bindless_supported));
-        slab_index
-    }
-
-    fn allocate(
-        &mut self,
-        fallback_images: &FallbackImage,
-        image_id: AssetId<Image>,
-    ) -> (LightmapSlabIndex, LightmapSlotIndex) {
-        let slab_index = match self.free_slabs.minimum() {
-            None => self.create_slab(fallback_images),
-            Some(slab_index) => slab_index.into(),
-        };
-
-        let slab = &mut self.slabs[usize::from(slab_index)];
-        let slot_index = slab.allocate(image_id);
-        if slab.is_full() {
-            self.free_slabs.remove(slab_index.into());
-        }
-
-        (slab_index, slot_index)
-    }
-
-    fn remove(
-        &mut self,
-        fallback_images: &FallbackImage,
-        slab_index: LightmapSlabIndex,
-        slot_index: LightmapSlotIndex,
-    ) {
-        let slab = &mut self.slabs[usize::from(slab_index)];
-        slab.remove(fallback_images, slot_index);
-
-        if !slab.is_full() {
-            self.free_slabs.grow_and_insert(slot_index.into());
-        }
-    }
 }
 
 impl LightmapSlab {
@@ -405,7 +269,6 @@ impl LightmapSlab {
             lightmaps: (0..count)
                 .map(|_| AllocatedLightmap {
                     gpu_image: fallback_images.d2.clone(),
-                    asset_id: None,
                 })
                 .collect(),
             free_slots_bitmask: (1 << count) - 1,
@@ -419,21 +282,19 @@ impl LightmapSlab {
     fn allocate(&mut self, image_id: AssetId<Image>) -> LightmapSlotIndex {
         let index = LightmapSlotIndex::from(self.free_slots_bitmask.trailing_zeros());
         self.free_slots_bitmask &= !(1 << u32::from(index));
-        self.lightmaps[usize::from(index)].asset_id = Some(image_id);
+        // self.lightmaps[usize::from(index)].asset_id = Some(image_id);
         index
     }
 
     fn insert(&mut self, index: LightmapSlotIndex, gpu_image: GpuImage) {
         self.lightmaps[usize::from(index)] = AllocatedLightmap {
             gpu_image,
-            asset_id: None,
         }
     }
 
     fn remove(&mut self, fallback_images: &FallbackImage, index: LightmapSlotIndex) {
         self.lightmaps[usize::from(index)] = AllocatedLightmap {
             gpu_image: fallback_images.d2.clone(),
-            asset_id: None,
         };
         self.free_slots_bitmask |= 1 << u32::from(index);
     }
@@ -470,50 +331,49 @@ impl LightmapSlab {
     }
 }
 
-impl From<u32> for LightmapSlabIndex {
-    fn from(value: u32) -> Self {
-        Self(NonMaxU32::new(value).unwrap())
+impl RenderLightmapsL {
+    /// Creates a new slab, appends it to the end of the list, and returns its
+    /// slab index.
+    fn create_slab(&mut self, other: &mut RenderLightmapsU, fallback_images: &FallbackImage) -> LightmapSlabIndex {
+        let slab_index = LightmapSlabIndex::from(self.slabs.len());
+        other.free_slabs.grow_and_insert(slab_index.into());
+        self.slabs
+            .push(LightmapSlab::new(fallback_images, other.bindless_supported));
+        slab_index
     }
-}
 
-impl From<usize> for LightmapSlabIndex {
-    fn from(value: usize) -> Self {
-        Self::from(value as u32)
+    fn allocate(
+        &mut self,
+        other: &mut RenderLightmapsU,
+        fallback_images: &FallbackImage,
+        image_id: AssetId<Image>,
+    ) -> (LightmapSlabIndex, LightmapSlotIndex) {
+        let slab_index = match other.free_slabs.minimum() {
+            None => self.create_slab(other, fallback_images),
+            Some(slab_index) => slab_index.into(),
+        };
+
+        let slab = &mut self.slabs[usize::from(slab_index)];
+        let slot_index = slab.allocate(image_id);
+        if slab.is_full() {
+            other.free_slabs.remove(slab_index.into());
+        }
+
+        (slab_index, slot_index)
     }
-}
 
-impl From<u32> for LightmapSlotIndex {
-    fn from(value: u32) -> Self {
-        Self(NonMaxU16::new(value as u16).unwrap())
-    }
-}
+    fn remove(
+        &mut self,
+        other: &mut RenderLightmapsU,
+        fallback_images: &FallbackImage,
+        slab_index: LightmapSlabIndex,
+        slot_index: LightmapSlotIndex,
+    ) {
+        let slab = &mut self.slabs[usize::from(slab_index)];
+        slab.remove(fallback_images, slot_index);
 
-impl From<usize> for LightmapSlotIndex {
-    fn from(value: usize) -> Self {
-        Self::from(value as u32)
-    }
-}
-
-impl From<LightmapSlabIndex> for usize {
-    fn from(value: LightmapSlabIndex) -> Self {
-        value.0.get() as usize
-    }
-}
-
-impl From<LightmapSlotIndex> for usize {
-    fn from(value: LightmapSlotIndex) -> Self {
-        value.0.get() as usize
-    }
-}
-
-impl From<LightmapSlotIndex> for u16 {
-    fn from(value: LightmapSlotIndex) -> Self {
-        value.0.get()
-    }
-}
-
-impl From<LightmapSlotIndex> for u32 {
-    fn from(value: LightmapSlotIndex) -> Self {
-        value.0.get() as u32
+        if !slab.is_full() {
+            other.free_slabs.grow_and_insert(slot_index.into());
+        }
     }
 }
