+++ crates/bevy_pbr/src/render/mod.rs
@@ -13,4 +13,4 @@ pub use light::*;
 pub use mesh::*;
 pub use mesh_view_bindings::*;
 pub use morph::*;
-pub use skin::{extract_skins, prepare_skins, skins_use_uniform_buffers, SkinUniforms, MAX_JOINTS};
+pub use skin::{extract_skins, prepare_skins, skins_use_uniform_buffers};
