+++ crates/bevy_render/src/render_resource/mod.rs
@@ -9,6 +9,7 @@ mod gpu_array_buffer;
 mod pipeline;
 mod pipeline_cache;
 mod pipeline_specializer;
+mod mesh_pipeline_specializer;
 pub mod resource_macros;
 mod specializer;
 mod storage_buffer;
@@ -25,6 +26,7 @@ pub use gpu_array_buffer::*;
 pub use pipeline::*;
 pub use pipeline_cache::*;
 pub use pipeline_specializer::*;
+pub use mesh_pipeline_specializer::*;
 pub use specializer::*;
 pub use storage_buffer::*;
 pub use texture::*;
